refactor and test for Color.alpha
